requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
accelerate==1.8.1
    # via
    #   -r requirements/dev.in
    #   peft
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via langchain
aiosignal==1.3.2
    # via aiohttp
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
bitsandbytes==0.46.0
    # via -r requirements/dev.in
boto3==1.38.41
    # via -r C:\workspace\AI\codecraft-ai\requirements\prod.in
botocore==1.38.41
    # via
    #   boto3
    #   s3transfer
certifi==2025.6.15
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
    #   uvicorn
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distlib==0.3.9
    # via virtualenv
faiss-cpu==1.11.0
    # via -r C:\workspace\AI\codecraft-ai\requirements\prod.in
fastapi==0.115.13
    # via -r C:\workspace\AI\codecraft-ai\requirements\prod.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   virtualenv
fonttools==4.58.4
    # via matplotlib
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.0
    # via
    #   -r requirements/dev.in
    #   huggingface-hub
    #   torch
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
huggingface-hub==0.33.0
    # via
    #   accelerate
    #   peft
    #   sentence-transformers
    #   tokenizers
    #   transformers
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.1
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
kiwisolver==1.4.8
    # via matplotlib
langchain==0.2.5
    # via -r C:\workspace\AI\codecraft-ai\requirements\prod.in
langchain-core==0.2.8
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.2.1
    # via
    #   -r C:\workspace\AI\codecraft-ai\requirements\prod.in
    #   langchain
langsmith==0.1.81
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via -r requirements/dev.in
mpmath==1.3.0
    # via sympy
multidict==6.5.0
    # via
    #   aiohttp
    #   yarl
narwhals==1.43.1
    # via plotly
networkx==3.5
    # via torch
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via
    #   -r C:\workspace\AI\codecraft-ai\requirements\prod.in
    #   accelerate
    #   bitsandbytes
    #   contourpy
    #   faiss-cpu
    #   langchain
    #   matplotlib
    #   pandas
    #   peft
    #   scikit-learn
    #   scipy
    #   transformers
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   accelerate
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   matplotlib
    #   peft
    #   plotly
    #   pytest
    #   transformers
pandas==2.3.0
    # via -r requirements/dev.in
peft==0.15.2
    # via -r requirements/dev.in
pillow==11.2.1
    # via
    #   matplotlib
    #   sentence-transformers
platformdirs==4.3.8
    # via virtualenv
plotly==6.1.2
    # via -r requirements/dev.in
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via -r requirements/dev.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via
    #   accelerate
    #   peft
pydantic==1.10.22
    # via
    #   -r C:\workspace\AI\codecraft-ai\requirements\prod.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
pygments==2.19.2
    # via pytest
pyparsing==3.2.3
    # via matplotlib
pytest==8.4.1
    # via -r requirements/dev.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via
    #   -r C:\workspace\AI\codecraft-ai\requirements\prod.in
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r C:\workspace\AI\codecraft-ai\requirements\prod.in
    #   accelerate
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   peft
    #   pre-commit
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   -r requirements/dev.in
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   transformers
ruff==0.12.0
    # via -r requirements/dev.in
s3transfer==0.13.0
    # via boto3
safetensors==0.5.3
    # via
    #   accelerate
    #   peft
    #   transformers
scikit-learn==1.7.0
    # via
    #   -r requirements/dev.in
    #   sentence-transformers
scipy==1.15.3
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==4.1.0
    # via -r requirements/dev.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via langchain
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
tabulate==0.9.0
    # via -r requirements/dev.in
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
torch==2.7.1
    # via
    #   -r requirements/dev.in
    #   accelerate
    #   bitsandbytes
    #   peft
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   peft
    #   sentence-transformers
    #   transformers
transformers==4.52.4
    # via
    #   -r requirements/dev.in
    #   peft
    #   sentence-transformers
typing-extensions==4.14.0
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   sentence-transformers
    #   sqlalchemy
    #   torch
tzdata==2025.2
    # via pandas
urllib3==1.26.20
    # via
    #   -r C:\workspace\AI\codecraft-ai\requirements\prod.in
    #   botocore
    #   requests
uvicorn[standard]==0.23.1
    # via -r C:\workspace\AI\codecraft-ai\requirements\prod.in
virtualenv==20.31.2
    # via pre-commit
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
